Add .gitattributes
.gitattributes contains support for dealing with line endings. Windows
stores line endings as CRLF where as on Unix systems only LF is used.
This creates issues when code is written and read on both systems.
This file tells git to resolve this issue by converting line endings to
a common standard in the repository but converting it on the user's
local machine.