Add fortran string utilities module
These utilities provide a string concatenation operator which allows string
concatenation with other types (integers, reals and bools). When the operator
is used to concatenate strings, trailing blanks are ignored (which is different
from the way Fortran implements it). This module is useful for building strings
containing multiple entities, e.g. for debug messages.