Add solver module and config file
The solver module provides subroutines that interface with the other modules
and can be used to propagate the state in time using a scheme.

The configuration file contains startup parameters that will be used by the
solver.

Currently, only the van Leer scheme can be used. More schemes are in the
pipeline. The choice of scheme will then be moved to the config file.