Add run script
The run script is a bash script that provides a hassle-free way to compile and
run the program. Always execute this run script to ensure the latest changes
are compiled.

The run script first compiles each of the modules individually, stopping
further compilation if errors are found in any module. The order of compilation
is the same as the order in which the modules are listed in the filelist
variable.

The run script also enables the logging feature; the script's output is sent to
a log file besides appearing on the screen. The program output is sent to the
log file only.