Add vtkwrite utility
For visualizing the output obtained from the solver, we use VisIt, an open
source, interactive, visualization tool. This tool accepts files written in the
vtk file format. It, however, was too onerous to generate the vtk file format
natively in fortran.  This python script converts the fvtk file format (as
generated by the fortran solver) into the vtk file format.