Add write fortran interface tool
This python script is a utility designed to generate fortran subroutine and
function interfaces when an explicit interface is required by another program
unit.

The purpose of this script was to be able to write a file for each method
containing the its interface. This file can then be include'd in the program
unit that requires the explicit interface. Later, if the subroutine is updated,
only the interface file needs to be regenerted and the update will be reflected
wherever the file is include'd. Of course the method call will have to be
updated manually!