Rename config.md to config.md.sample
Currently git is tracking the config file (config.md). However, every time a
change is made to the file, git wants to update it on the repository. This is
not desired as multiple test cases need to be run but the fields in the file
remain the same.

Once this file is renamed, .gitignore will be updated so that git stops
tracking config.md and instead tracks only config.md.sample.