Don't use resnorm_previous for convergence check
resnorm refers to the 2-norm of the residue. resnorm_previous refers to the
norm of the residue in the previous iteration. It seems like the residue will
not change much in one iteration and hence testing for this change in the
convergence check seems unprofitable.

Now, convergence is checked only by testing for the change in the current
residue norm compared with the norm of the residue when the solver began.