Write output to a temp file first (then rename)
The daemon to translate and visualize the output files of the solver listens
for the creation of new files. When a new file is found, the daemon converts
the program output (*.fvtk) to *.vtk files. These files are sent to VisIt for
visualization through the python script (visit -cli mode).

The daemon listens for files of a particular name (output*.fvtk). This means
that as soon as a *.fvtk is created, the daemon tries to translate it. This may
happen even if the file writing operation is still underway. To prevent this
behavior, the data is initially written to a temp file first
(output*.fvtk.part). Once the write is complete, the file is renamed to the
correct filename (output*.fvtk).

A similar change was done in vtkwrite.py which reads in the output*.fvtk and
effectively outputs output*.vtk.