Add daemon to translate and visualize output files
This daemon can be used to listen for files being output by the solver (*.fvtk)
and translate them to a format readable by VisIt (*.vtk). It also sends these
translated *.vtk files to VisIt (through the -cli mode) for visualization. This
enables visualization of program output in tandem with the computation itself.