Reimplement LDFSS(0) to make use of Van-Leer
The working of LDFSS(0) is very similar to Van-Leer except for a small change
in the computation of the numerical speeds. It is, hence, fitting for the
implementation of LDFSS(0) to take advantage of the implementation of Van-Leer.
This commit effects this change from the naive implementation of a stand-alone
LDFSS(0) module to one that depends on the implementation of the Van-Leer
scheme.