Decouple scheme selection from solver
Until now, the solver was responsible for invoking the correct scheme based on
the input supplied to the user. This was making the solver code very cluttered
(even though the code was physically stored in separate include files).

The new implementation provides a new scheme module which is responsible for
the aforementioned book keeping, thus making the code a bit more modular. Now
even the scheme name and residue are managed by the scheme module.