Update Van Leer to use left and right states
The previous implementation of Van Leer was using a first order space
reconstruction. Now, with the addition of the face_interpolant module, we can
rewrite the implementation of Van Leer in terms of the face reconstructed data
provided by that module. This allows us to achieve higher order accurate space
formulations by changing the scheme used to reconstruct the face interpolated
values.