Remove outdated sound speed computations
With the addition of the face_interpolant module, the values of the variables
at the faces are being reconstructed there. Hence, the computation of sound
speed at the faces (which is used in many places) has been moved to that module
making the previous code responsible for this outdated. This commit removes all
this unnecessary code.