Add PPM (Piecewise Parabolic Method)
PPM is a method used to get face interpolated data using cell averaged data.
This can be applied on top of a scheme to obtain a scheme which is higher order
accurate in space. This commit introduces the code for PPM, an option in the
config file to enable the user to choose whether or not PPM should be used as
well as the accompanying interface to effect this change in method (in the
face_interpolant module).