Create checkpointing mechanism
A field has been added to the config file where the user can enter after how
many iterations a checkpoint should be created. With the current
implementation, an initial checkpoint is created (during solver setup,
right after the initialization) and then every time the fixed number of
iterations are completed.

The earlier method to do this was to hardcode the value in the main file; it
was not a feature of the solver. This meant that the code had to be recompiled
every time a change had to be made. This recompilation step is not required
anymore.