Allow fvtk's to contain comments
When dealing with global time steps, it sometimes makes sense to know what the
time is on the simulation clock. The second line in vtk files is meant for
comments. This motivated the idea to use that comment line to print the system
time.

This commit allows the writestate subroutine to receive an arbitrary comment
which it then prints as the first line of the fvtk file (if a comment is not
provided, the default is written). The vtkwrite python utility has been
upgraded to be able to read the new fvtk format (the only change being the
extra comment line in the beginning).

The vtkwrite utility has also been revamped a bit to make the code a bit
cleaner.